* master: (113 commits)
  MBS-3451, fix unterminated <a> link.
  MBS-1169: Add more HTML sanitization to "edit medium" edit template
  Fix processing bug with cover art script
  MBS-3643: Allow musixmatch for the lyrics whitelist
  Add IMSLP to the list of URL specializations
  More space between the two link sections
  Fix quotation marks on tag lookup results
  Properly catch NoChanges, but still die on other exceptions.
  Display release group ARs on the release relationships page
  Display release group URLs on release pages
  Split the arranger AR into generic/vocal/instrument ARs
  Trying to fix the YouTube stuff 2 (with help by warp)
  Trying to fix YouTube cleanup
  Correctly update artist_credit.name on artist merge
  Use Try::Tiny in Controller::AutoEditorElections.
  MBS-3719 Fix display issues with search results containing standalone recordings - alternating row colouring broken for standalone recordings - no tagger links for standalone recordings - columns not highlighted correctly for standalone recordings
  Make tagger links for recordings bring the release down into Picard; not a NAT/standalone recording.
  Replace all use of TryCatch with TryTiny.
  MBS-3522, rename inaccurate "@comment_required" to "@new_entities".
  MBS-3522, remove unneccesary musicbrainz_unaccent function.
  ...

+++ lib/MusicBrainz/Server/Controller/FreeDB.pm
@@ -3,7 +3,7 @@ package MusicBrainz::Server::Controller::FreeDB;
 use Moose;
 use MusicBrainz::Server::Translation qw( l );
 use MusicBrainz::Server::Validation qw( is_freedb_id );
-use TryCatch;
+use Try::Tiny;
 
 BEGIN { extends 'MusicBrainz::Server::Controller' }
 
@@ -21,9 +21,11 @@ sub _load {
         try {
             return $c->model('FreeDB')->lookup($category, $id);
         }
-        catch (MusicBrainz::Server::Exceptions::InvalidInput $e) {
-            $c->stash( message => $e->message );
-            $c->detach('/error_500');
+        catch {
+            if (ref($_) eq 'MusicBrainz::Server::Exceptions::InvalidInput') {
+                $c->stash( message => $_->message );
+                $c->detach('/error_500');
+            }
         };
     }
     else {
